Summer_workshop_TEAM_2_BACKEND/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.7.0
python-dotenv==1.1.1
google-generativeai==0.8.5
python-multipart==0.0.20
requests==2.32.4
typing-extensions==4.14.1
